requirements.txt
CTkMessagebox==2.4
CTkMessagebox==2.5
CTkToolTip==0.8
CTkToolTip==0.8
customtkinter==5.1.2
keyboard==0.13.5
matplotlib==3.7.2
mouse==0.7.1
mss==8.0.3
networkx==3.1
numpy==1.23.5
opencv_contrib_python==4.6.0.66
opencv_python==4.5.5.64
opencv_python_headless==4.5.4.60
pandas==2.0.0
Pillow==10.0.0
pycryptodome==3.17
PyGetWindow==0.0.9
pynput==1.7.6
tesserocr==2.5.2
